a bit of debug statements
